version 1.35
Release Notes
Factory default of Bluetooth is changed from Off to On. 
Laser Atlanta Stealth gun detection is added.
K block 3 filter (24.123~24.124GHz) is added.
User K Block Filter 1,2,3,4,5 are added.
Selective mute memory frequency feature is added.
Auto Mute Alert Level feature is added. The existing Auto Mute On/Off is
deleted from the menu and Auto Mute is integrated into Auto Mute Alert Level as below.
Off is the same as the behavior of the existing Auto Mute Off, and the Auto mute function does not work when setting it to off.
When setting it to Off, the Auto Mute Level function is not turned off, but the Auto Mute function is turned off.
If setting it to 1~4 level, the alert is auto muted only when the alert level is below the set Auto Mute level,
and if an alert level exceeds the set Auto Mute level, it will not be auto muted.
If setting it to 5 level, Auto mute is applied to all levels, so it works the same as the existing Auto Mute On.
(Auto Mute Level: 5 = Auto Mute On)
Alert temporary volume feature is added. During an alert, it works separately and temporarily from the main volume.
Rear Balance function for each band (K, Ka, X) is added.
Main Volume setting is added to the Menu.
The Bluetooth On/Off and BT Pairing are moved to be before GPS On/Off menu in the Menu.
Low speed (5Km/h) filter for POI (Redlight, Speed cam) alerts is added.
The Auto Mute Memory related values are adjusted to Delete count 5, Visit interval 4, Delete radius 60m, Mute radius 400m, Delete check 2sec.
GPS Error1 is fixed.
The audio announcement for DST ON/OFF reversed in the menu is fixed.